+++ Sources/Nuke/ImageContainer.swift
@@ -95,9 +95,6 @@ public struct ImageContainer: @unchecked Sendable {
             self.rawValue = value
         }
 
-        // For internal purposes.
-        static let isThumbnailKey: UserInfoKey = "com.github/kean/nuke/skip-decompression"
-
         /// A user info key to get the scan number (Int).
         public static let scanNumberKey: UserInfoKey = "com.github/kean/nuke/scan-number"
     }