+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 ## Nuke (WIP)
 
 - Add missing `@ImagePipelineActor` isolation to the new `willLoadData` method in `ImagePipeline.Delegate` to avoid thread hops
+- Minor other performance improvements 
 
 ## Nuke 13.0.3